requirements.txt
matplotlib
numpy
pandas
pathlib
Pillow
scikit-learn
scipy
sklearn
torch
torchvision
tqdm
yacs
efficientnet_pytorch
